Refatoração

+++ checkpoint2/checkpoint2/src/main/java/br/com/fiap/checkpoint2/repositories/IGenericRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface IGenericRepository<T> extends JpaRepository<T, Integer>{
+public interface IGenericRepository<T> extends JpaRepository<T, Long>{
 
 }
